backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
openai==1.3.7
requests==2.31.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
returns==0.21.0
gunicorn==21.2.0
